backend/requirements.txt
fastapi
uvicorn[standard]==0.30.0
python-dotenv
pydantic~=2.11.7
starlette~=0.37.2

# database
SQLAlchemy
psycopg2-binary==2.9.10
psycopg
pgvector

# HTTP & utility
httpx==0.27.0
itsdangerous==2.2.0
requests==2.32.5
tqdm==4.67.1

# Security / Authentication
authlib==1.6.4
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography>=40.0.0,<50.0.0
# Automatic Payment
APScheduler==3.10.4

# TensorFlow / Keras
tensorflow==2.15.0
tf-keras

# Core NLP Libraries
transformers==4.57.1
peft==0.17.1
# ▲ T5 Model Training Environment
sentence-transformers==2.7.0
huggingface-hub>=0.34.0,<1.0
accelerate==1.2.1
datasets==3.2.0
sentencepiece==0.2.0
codecarbon==2.8.2

# Korean NLP
kiwipiepy
konlpy
nltk==3.8.1

# Audio Processing
librosa==0.10.1
soxr

# Other NLP/ML
gensim==4.3.2

# Game Implementation
hgtk==0.2.1

# GitHub Direct Installation (Install Last)
git+https://github.com/ssut/py-hanspell.git@master
# git+https://github.com/haven-jeon/PyKoSpacing.git@master  # TFSMLayer error

# Deployment Related
gunicorn==20.1.0

# Installation Order
# 0. pip install --upgrade pip

# 1. TTS Library = Install Melo
# git clone https://github.com/myshell-ai/MeloTTS.git
# cd MeloTTS
# pip install -e .
# python -m unidic download

# $env:PYTHONUTF8=1
# 2. pip install --no-cache-dir -r requirements.txt

